+++ darts/models/forecasting/croston.py
@@ -66,7 +66,8 @@ class Croston(FutureCovariatesLocalForecastingModel):
                     'datetime_attribute': {'future': ['hour', 'dayofweek']},
                     'position': {'future': ['relative']},
                     'custom': {'future': [encode_year]},
-                    'transformer': Scaler()
+                    'transformer': Scaler(),
+                    'tz': 'CET'
                 }
             ..