+++ darts/models/forecasting/linear_regression_model.py
@@ -94,7 +94,8 @@ class LinearRegressionModel(RegressionModel, _LikelihoodMixin):
                     'datetime_attribute': {'future': ['hour', 'dayofweek']},
                     'position': {'past': ['relative'], 'future': ['relative']},
                     'custom': {'past': [encode_year]},
-                    'transformer': Scaler()
+                    'transformer': Scaler(),
+                    'tz': 'CET'
                 }
             ..
         likelihood
@@ -124,6 +125,7 @@ class LinearRegressionModel(RegressionModel, _LikelihoodMixin):
         Examples
         --------
         Deterministic forecasting, using past/future covariates (optional)
+
         >>> from darts.datasets import WeatherDataset
         >>> from darts.models import LinearRegressionModel
         >>> series = WeatherDataset().load()