+++ darts/utils/data/tabularization.py
@@ -883,12 +883,27 @@ def _create_lagged_data_by_moving_window(
             # for all feature times - these values will become labels.
             # If `start_time` not included in `time_index_i`, can 'manually' calculate
             # what its index *would* be if `time_index_i` were extended to include that time:
-            if not is_target_series and (time_index_i[-1] <= start_time):
-                start_time_idx = (
-                    len(time_index_i)
-                    - 1
-                    + (start_time - time_index_i[-1]) // series_i.freq
-                )
+            if not is_target_series and (time_index_i[-1] < start_time):
+                # Series frequency represents a non-ambiguous timedelta value (not ‘M’, ‘Y’ or ‘y’)
+                if pd.to_timedelta(series_i.freq, errors="coerce") is not pd.NaT:
+                    start_time_idx = (
+                        len(time_index_i)
+                        - 1
+                        + (start_time - time_index_i[-1]) // series_i.freq
+                    )
+                else:
+                    # Create a temporary DatetimeIndex to extract the actual start index.
+                    start_time_idx = (
+                        len(time_index_i)
+                        - 1
+                        + len(
+                            pd.date_range(
+                                start=time_index_i[-1] + series_i.freq,
+                                end=start_time,
+                                freq=series_i.freq,
+                            )
+                        )
+                    )
             elif not is_target_series and (time_index_i[0] >= start_time):
                 start_time_idx = max_lag_i
             # If `start_time` *is* included in `time_index_i`, need to binary search `time_index_i`