+++ darts/utils/historical_forecasts/utils.py
@@ -1,5 +1,5 @@
 from types import SimpleNamespace
-from typing import Any, Callable, Optional, Tuple, Union
+from typing import Any, Callable, Optional, Sequence, Tuple, Union
 
 try:
     from typing import Literal
@@ -8,6 +8,7 @@ except ImportError:
 
 import numpy as np
 import pandas as pd
+from numpy.typing import ArrayLike
 
 from darts.logging import get_logger, raise_if_not, raise_log
 from darts.timeseries import TimeSeries
@@ -27,6 +28,7 @@ TimeIndex = Union[
 def _historical_forecasts_general_checks(model, series, kwargs):
     """
     Performs checks common to ForecastingModel and RegressionModel backtest() methods
+
     Parameters
     ----------
     model
@@ -241,6 +243,8 @@ def _historical_forecasts_start_warnings(
 def _get_historical_forecastable_time_index(
     model,
     series: TimeSeries,
+    forecast_horizon: int,
+    overlap_end: bool,
     past_covariates: Optional[TimeSeries] = None,
     future_covariates: Optional[TimeSeries] = None,
     is_training: Optional[bool] = False,
@@ -264,11 +268,21 @@ def _get_historical_forecastable_time_index(
         timestamp is a timestamp that has a training sample of length at least ``self.training_sample_length``
             preceding it.
 
+    Additionally, it accounts for auto-regression (forecast_horizon > model.output_chunk_length , and overlap_end.
+        - In case of auto-regression, we have to step back the last possible predictable/trainable time step if the
+          covariates are too short
+        - In case overlap_end=False, the latest possible predictable/trainable time step is shifted back if a
+          prediction starting from that point would go beyond the end of `series`.
+
 
     Parameters
     ----------
     series
         A target series.
+    forecast_horizon
+        The forecast horizon for the predictions.
+    overlap_end
+        Whether the returned forecasts can go beyond the series' end or not.
     past_covariates
         Optionally, a past covariates.
     future_covariates
@@ -288,26 +302,35 @@ def _get_historical_forecastable_time_index(
     >>> model = LinearRegressionModel(lags=3, output_chunk_length=2)
     >>> model.fit(train_series)
     >>> series = TimeSeries.from_times_and_values(pd.date_range('2000-01-01', '2000-01-10'), np.arange(10))
-    >>> model._get_historical_forecastable_time_index(series=series, is_training=False)
-    DatetimeIndex(['2000-01-04', '2000-01-05', '2000-01-06', '2000-01-07',
-            '2000-01-08', '2000-01-09', '2000-01-10'],
-            dtype='datetime64[ns]', freq='D')
+    >>> model._get_historical_forecastable_time_index(series=series, is_training=False, forecast_horizon=1)
+    DatetimeIndex(
+            ['2000-01-04', '2000-01-05', '2000-01-06', '2000-01-07', '2000-01-08', '2000-01-09', '2000-01-10'],
+            dtype='datetime64[ns]', freq='D'
+    )
     >>> model._get_historical_forecastable_time_index(series=series, is_training=True)
-    DatetimeIndex(['2000-01-06', '2000-01-08', '2000-01-09', '2000-01-10'],
-            dtype='datetime64[ns]', freq='D')
-
+    DatetimeIndex(['2000-01-06', '2000-01-08', '2000-01-09', '2000-01-10'], dtype='datetime64[ns]', freq='D')
     >>> model = NBEATSModel(input_chunk_length=3, output_chunk_length=3)
     >>> model.fit(train_series, train_past_covariates)
     >>> series = TimeSeries.from_times_and_values(pd.date_range('2000-10-01', '2000-10-09'), np.arange(8))
-    >>> past_covariates = TimeSeries.from_times_and_values(pd.date_range('2000-10-03', '2000-10-20'),
-    np.arange(18))
-    >>> model._get_historical_forecastable_time_index(series=series, past_covariates=past_covariates,
-    is_training=False)
+    >>> past_covariates = TimeSeries.from_times_and_values(
+    >>>     pd.date_range('2000-10-03', '2000-10-20'),
+    >>>     np.arange(18)
+    >>> )
+    >>> model._get_historical_forecastable_time_index(
+    >>>     series=series,
+    >>>     past_covariates=past_covariates,
+    >>>     is_training=False,
+    >>>     forecast_horizon=1,
+    >>> )
     DatetimeIndex(['2000-10-06', '2000-10-07', '2000-10-08', '2000-10-09'], dtype='datetime64[ns]', freq='D')
-    >>> model._get_historical_forecastable_time_index(series=series, past_covariates=past_covariates,
-    is_training=True)
-    DatetimeIndex(['2000-10-09'], dtype='datetime64[ns]', freq='D') # Only one point is trainable, and
-    # corresponds to the first point after we reach a common subset of timestamps of training_sample_length length.
+    >>>  # Only one point is trainable; it corresponds to the first point after we reach a common subset of
+    >>> # timestamps of training_sample_length length.
+    >>> model._get_historical_forecastable_time_index(
+    >>>     series=series,
+    >>>     past_covariates=past_covariates,
+    >>>     is_training=True,
+    >>> )
+    DatetimeIndex(['2000-10-09'], dtype='datetime64[ns]', freq='D')
     """
 
     (
@@ -319,59 +342,76 @@ def _get_historical_forecastable_time_index(
         max_future_cov_lag,
     ) = model.extreme_lags
 
+    # max_target_lag < 0 are local models which can predict for n (horizon) -> infinity (no auto-regression)
+    is_autoregression = max_target_lag >= 0 and forecast_horizon > max_target_lag + 1
+
     if min_target_lag is None:
         min_target_lag = 0
 
     # longest possible time index for target
-    start = (
-        series.start_time() + (max_target_lag - min_target_lag + 1) * series.freq
-        if is_training
-        else series.start_time() - min_target_lag * series.freq
-    )
+    if is_training:
+        start = (
+            series.start_time() + (max_target_lag - min_target_lag + 1) * series.freq
+        )
+    else:
+        start = series.start_time() - min_target_lag * series.freq
     end = series.end_time() + 1 * series.freq
 
     intersect_ = (start, end)
 
     # longest possible time index for past covariates
     if (min_past_cov_lag is not None) and (past_covariates is not None):
-        start_pc = (
-            past_covariates.start_time()
-            - (min_past_cov_lag - max_target_lag - 1) * past_covariates.freq
-            if is_training
-            else past_covariates.start_time() - min_past_cov_lag * past_covariates.freq
-        )
-        end_pc = past_covariates.end_time() - max_past_cov_lag * past_covariates.freq
-        tmp_ = (start_pc, end_pc)
-
-        if intersect_ is not None:
-            intersect_ = (
-                max([intersect_[0], tmp_[0]]),
-                min([intersect_[1], tmp_[1]]),
+        if is_training:
+            start_pc = (
+                past_covariates.start_time()
+                - (min_past_cov_lag - max_target_lag - 1) * past_covariates.freq
             )
         else:
-            intersect_ = tmp_
+            start_pc = (
+                past_covariates.start_time() - min_past_cov_lag * past_covariates.freq
+            )
+
+        shift_pc_end = max_past_cov_lag
+        if is_autoregression:
+            # we step back in case of auto-regression
+            shift_pc_end += forecast_horizon - (max_target_lag + 1)
+        end_pc = past_covariates.end_time() - shift_pc_end * past_covariates.freq
+
+        intersect_ = (
+            max([intersect_[0], start_pc]),
+            min([intersect_[1], end_pc]),
+        )
 
     # longest possible time index for future covariates
     if (min_future_cov_lag is not None) and (future_covariates is not None):
-        start_fc = (
-            future_covariates.start_time()
-            - (min_future_cov_lag - max_target_lag - 1) * future_covariates.freq
-            if is_training
-            else future_covariates.start_time()
-            - min_future_cov_lag * future_covariates.freq
-        )
-        end_fc = (
-            future_covariates.end_time() - max_future_cov_lag * future_covariates.freq
-        )
-        tmp_ = (start_fc, end_fc)
-
-        if intersect_ is not None:
-            intersect_ = (
-                max([intersect_[0], tmp_[0]]),
-                min([intersect_[1], tmp_[1]]),
+        if is_training:
+            start_fc = (
+                future_covariates.start_time()
+                - (min_future_cov_lag - max_target_lag - 1) * future_covariates.freq
             )
         else:
-            intersect_ = tmp_
+            start_fc = (
+                future_covariates.start_time()
+                - min_future_cov_lag * future_covariates.freq
+            )
+
+        shift_fc_end = max_future_cov_lag
+        if is_autoregression:
+            # we step back in case of auto-regression
+            shift_fc_end += forecast_horizon - (max_target_lag + 1)
+        end_fc = future_covariates.end_time() - shift_fc_end * future_covariates.freq
+
+        intersect_ = (
+            max([intersect_[0], start_fc]),
+            min([intersect_[1], end_fc]),
+        )
+
+    # overlap_end = True -> predictions must not go beyond end of target series
+    if (
+        not overlap_end
+        and intersect_[1] + (forecast_horizon - 1) * series.freq > series.end_time()
+    ):
+        intersect_ = (intersect_[0], end - forecast_horizon * series.freq)
 
     # end comes before the start
     if intersect_[1] < intersect_[0]:
@@ -395,12 +435,9 @@ def _get_historical_forecastable_time_index(
 
 
 def _adjust_historical_forecasts_time_index(
-    model,
     series: TimeSeries,
     series_idx: int,
     historical_forecasts_time_index: TimeIndex,
-    forecast_horizon: int,
-    overlap_end: bool,
     start: Optional[Union[pd.Timestamp, float, int]],
     start_format: Literal["position", "value"],
     show_warnings: bool,
@@ -409,19 +446,6 @@ def _adjust_historical_forecasts_time_index(
     Shrink the beginning and end of the historical forecasts time index based on the values of `start`,
     `forecast_horizon` and `overlap_end`.
     """
-    # shift the end of the forecastable index based on `overlap_end`` and `forecast_horizon``
-    last_valid_pred_time = model._get_last_prediction_time(
-        series,
-        forecast_horizon,
-        overlap_end,
-        latest_possible_prediction_start=historical_forecasts_time_index[-1],
-    )
-
-    historical_forecasts_time_index = (
-        historical_forecasts_time_index[0],
-        min(historical_forecasts_time_index[-1], last_valid_pred_time),
-    )
-
     # when applicable, shift the start of the forecastable index based on `start`
     if start is not None:
         if start_format == "value":
@@ -456,13 +480,17 @@ def _get_historical_forecast_predict_index(
     series_idx: int,
     past_covariates: Optional[TimeSeries],
     future_covariates: Optional[TimeSeries],
+    forecast_horizon: int,
+    overlap_end: bool,
 ) -> TimeIndex:
     """Obtain the boundaries of the predictable time indices, raise an exception if None"""
     historical_forecasts_time_index = _get_historical_forecastable_time_index(
-        model,
-        series,
-        past_covariates,
-        future_covariates,
+        model=model,
+        series=series,
+        forecast_horizon=forecast_horizon,
+        overlap_end=overlap_end,
+        past_covariates=past_covariates,
+        future_covariates=future_covariates,
         is_training=False,
         reduce_to_bounds=True,
     )
@@ -486,16 +514,20 @@ def _get_historical_forecast_train_index(
     series_idx: int,
     past_covariates: Optional[TimeSeries],
     future_covariates: Optional[TimeSeries],
+    forecast_horizon: int,
+    overlap_end: bool,
 ) -> TimeIndex:
     """
     Obtain the boundaries of the time indices usable for training, raise an exception if training is required and
     no indices are available.
     """
     historical_forecasts_time_index = _get_historical_forecastable_time_index(
-        model,
-        series,
-        past_covariates,
-        future_covariates,
+        model=model,
+        series=series,
+        forecast_horizon=forecast_horizon,
+        overlap_end=overlap_end,
+        past_covariates=past_covariates,
+        future_covariates=future_covariates,
         is_training=True,
         reduce_to_bounds=True,
     )
@@ -598,17 +630,20 @@ def _get_historical_forecast_boundaries(
     """
     # obtain forecastable indexes boundaries, as values from the time index
     historical_forecasts_time_index = _get_historical_forecast_predict_index(
-        model, series, series_idx, past_covariates, future_covariates
+        model,
+        series,
+        series_idx,
+        past_covariates,
+        future_covariates,
+        forecast_horizon,
+        overlap_end,
     )
 
     # adjust boundaries based on start, forecast_horizon and overlap_end
     historical_forecasts_time_index = _adjust_historical_forecasts_time_index(
-        model=model,
         series=series,
         series_idx=series_idx,
         historical_forecasts_time_index=historical_forecasts_time_index,
-        forecast_horizon=forecast_horizon,
-        overlap_end=overlap_end,
         start=start,
         start_format=start_format,
         show_warnings=show_warnings,
@@ -646,10 +681,16 @@ def _get_historical_forecast_boundaries(
     if series.has_range_index:
         hist_fct_tgt_start = series.get_index_at_point(hist_fct_tgt_start)
         hist_fct_tgt_end = series.get_index_at_point(hist_fct_tgt_end) + 1
-        hist_fct_pc_start = series.get_index_at_point(hist_fct_pc_start)
-        hist_fct_pc_end = series.get_index_at_point(hist_fct_pc_end) + 1
-        hist_fct_fc_start = series.get_index_at_point(hist_fct_fc_start)
-        hist_fct_fc_end = series.get_index_at_point(hist_fct_fc_end) + 1
+        if past_covariates is not None:
+            hist_fct_pc_start = past_covariates.get_index_at_point(hist_fct_pc_start)
+            hist_fct_pc_end = past_covariates.get_index_at_point(hist_fct_pc_end) + 1
+        else:
+            hist_fct_pc_start, hist_fct_pc_end = None, None
+        if future_covariates is not None:
+            hist_fct_fc_start = future_covariates.get_index_at_point(hist_fct_fc_start)
+            hist_fct_fc_end = future_covariates.get_index_at_point(hist_fct_fc_end) + 1
+        else:
+            hist_fct_fc_start, hist_fct_fc_end = None, None
 
     return (
         historical_forecasts_time_index[0],
@@ -661,3 +702,113 @@ def _get_historical_forecast_boundaries(
         hist_fct_fc_start,
         hist_fct_fc_end,
     )
+
+
+def _check_optimizable_historical_forecasts_global_models(
+    model,
+    forecast_horizon: int,
+    retrain: Union[bool, int, Callable[..., bool]],
+    show_warnings: bool,
+    allow_autoregression: bool,
+) -> bool:
+    """
+    Historical forecast can be optimized only if `retrain=False`. If `allow_autoregression=False`, historical forecasts
+    can be optimized only if `forecast_horizon <= model.output_chunk_length` (no auto-regression required).
+    """
+
+    retrain_off = (retrain is False) or (retrain == 0)
+    is_autoregressive = forecast_horizon > model.output_chunk_length
+    if retrain_off and (
+        not is_autoregressive or (is_autoregressive and allow_autoregression)
+    ):
+        return True
+
+    if show_warnings:
+        if not retrain_off:
+            logger.warning(
+                "`enable_optimization=True` is ignored because `retrain` is not `False` or `0`."
+                "To hide this warning, set `show_warnings=False` or `enable_optimization=False`."
+            )
+        if is_autoregressive:
+            logger.warning(
+                "`enable_optimization=True` is ignored because `forecast_horizon > model.output_chunk_length`."
+                "To hide this warning, set `show_warnings=False` or `enable_optimization=False`."
+            )
+
+    return False
+
+
+def _process_historical_forecast_input(
+    model,
+    series: Optional[Sequence[TimeSeries]],
+    past_covariates: Optional[Sequence[TimeSeries]] = None,
+    future_covariates: Optional[Sequence[TimeSeries]] = None,
+    forecast_horizon: int = 1,
+    allow_autoregression: bool = False,
+):
+    if not model._fit_called:
+        raise_log(
+            ValueError("Model has not been fit yet."),
+            logger,
+        )
+
+    if not allow_autoregression and forecast_horizon > model.output_chunk_length:
+        raise_log(
+            ValueError(
+                "`forecast_horizon > model.output_chunk_length` requires auto-regression which is not "
+                "supported in this optimized routine."
+            ),
+            logger,
+        )
+
+    # manage covariates, usually handled by RegressionModel.predict()
+    if past_covariates is None and model.past_covariate_series is not None:
+        past_covariates = [model.past_covariate_series] * len(series)
+    if future_covariates is None and model.future_covariate_series is not None:
+        future_covariates = [model.future_covariate_series] * len(series)
+
+    model._verify_static_covariates(series[0].static_covariates)
+
+    if model.encoders.encoding_available:
+        past_covariates, future_covariates = model.generate_fit_predict_encodings(
+            n=forecast_horizon,
+            series=series,
+            past_covariates=past_covariates,
+            future_covariates=future_covariates,
+        )
+    return series, past_covariates, future_covariates
+
+
+def _process_predict_start_points_bounds(
+    series: Sequence[TimeSeries], bounds: ArrayLike, stride: int
+) -> Tuple[np.ndarray, np.ndarray]:
+    """Processes the historical forecastable time index bounds (earliest, and latest possible prediction
+    start points).
+
+    Parameters
+    ----------
+    bounds
+        An array of shape (n series, 2), with the left and right prediction start point bounds per series.
+    stride
+        The number of time steps between two consecutive predictions.
+
+    Returns
+    -------
+    (np.ndarray, np.ndarray)
+        The adjusted bounds: the right bounds are adjusted to be a multiple of 'stride' ahead of the left bounds.
+        The number of resulting predicted series per input series respecting stride and bounds.
+    """
+    bounds = bounds if isinstance(bounds, np.ndarray) else np.array(bounds)
+    if not bounds.shape == (len(series), 2):
+        raise_log(
+            ValueError(
+                "`bounds` must be an array like with shape `(n target series, 2)`, "
+                "with the start and end bounds of each series"
+            ),
+            logger=logger,
+        )
+    # we might have some steps that are too long considering stride
+    steps_too_long = (bounds[:, 1] - bounds[:, 0]) % stride
+    bounds[:, 1] -= steps_too_long
+    cum_lengths = np.cumsum(np.diff(bounds) // stride + 1)
+    return bounds, cum_lengths